[core] remove extra HashSet<T> from PropertyMapper
I setup a new benchmark for `PropertyMapper` with 50 properties,
and sending a single update. Reviewing `PropertyMapper`, I found an
extra `HashSet<T>` of the keys -- that doesn't appear to be used
anywhere. It just seems like we create a copy of every
`PropertyMapper` key on startup or when properties change?

Removing the `HashSet<T>` improved the benchmark by about ~33%, and
memory usage by ~23%:

|             Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| --UpdateProperties | 5.915 us | 0.1108 us | 0.1232 us | 2.0294 | 0.0839 |     17 KB |
| ++UpdateProperties | 3.924 us | 0.0385 us | 0.0360 us | 1.6098 | 0.0610 |     13 KB |

This should improve dotnet/maui startup, there are static
`PropertyMapper`s on each handler type. It also should improve general
property changes, as it appeared that we were clearing and recreating
the `HashSet<T>` when properties change.

I did not see a noticeable difference in a `dotnet new maui` startup on
Android. (The changes seemed to just be within margin of error)

This removes the APIs from `Microsoft.Maui.PropertyMapper`:

    protected HashSet<string> PopulateKeys(ref HashSet<string>? returnList)
    protected virtual void ClearKeyCache()
    public virtual IReadOnlyCollection<string> UpdateKeys { get; }

I did not find usage of these APIs, so it seems like they are not needed?