[ios] fix memory leak with `Page` + `Layout`
Further fixes: https://github.com/dotnet/maui/issues/13520

7d0af63c did not appear to be a complete fix for #13520, as when I
retested the customer sample, I found I needed to remove a
`VerticalStackLayout` for the problem to go away.

I could reproduce the issue in a test by doing:

    var page = new ContentPage { Title = "Page 2", Content = new VerticalStackLayout { new Label() } };

After lots of debugging, I found the underlying issue was that
`MauiView` on iOS held a reference to the cross-platform `IView`.

After changing the backing field for this to be a `WeakReference<IView>`
the above problem appears to be solved.

I also removed `LayoutView.CrossplatformArrange/Measure` as we can just
use the `View` property directly instead.